Add all team members to OWNERS file
ClusterFuzz uses this file to determine the CC list for new bugs.
If ClusterFuzz thinks a bug is scary, it'll restrict visibility to the assignee and people CC'd on the issue.
As we go into OT, we want the whole team to be able to see these issues, so expand the OWNERS to the whole team.

Change-Id: Idf717eb5f09c1712ceb4cf58174c59c4032fa5f6
Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/61061
Reviewed-by: David Neto <dneto@google.com>
Commit-Queue: Ben Clayton <bclayton@google.com>
Kokoro: Ben Clayton <bclayton@google.com>